+++ lib/sidekiq/job/iterable.rb
@@ -30,6 +30,40 @@ module Sidekiq
         @_cursor = nil
         @_start_time = nil
         @_runtime = 0
+        @_args = nil
+        @_cancelled = nil
+      end
+
+      def arguments
+        @_args
+      end
+
+      # Three days is the longest period you generally need to wait for a retry to
+      # execute when using the default retry scheme. We don't want to "forget" the job
+      # is cancelled before it has a chance to execute and cancel itself.
+      CANCELLATION_PERIOD = (3 * 86_400).to_s
+
+      # Set a flag in Redis to mark this job as cancelled.
+      # Cancellation is asynchronous and is checked at the start of iteration
+      # and every 5 seconds thereafter as part of the recurring state flush.
+      def cancel!
+        return @_cancelled if cancelled?
+
+        key = "it-#{jid}"
+        _, result, _ = Sidekiq.redis do |c|
+          c.pipelined do |p|
+            p.hsetnx(key, "cancelled", Time.now.to_i)
+            p.hget(key, "cancelled")
+            # TODO When Redis 7.2 is required
+            # p.expire(key, Sidekiq::Job::Iterable::STATE_TTL, "nx")
+            p.expire(key, Sidekiq::Job::Iterable::STATE_TTL)
+          end
+        end
+        @_cancelled = result.to_i
+      end
+
+      def cancelled?
+        @_cancelled
       end
 
       # A hook to override that will be called when the job starts iterating.
@@ -91,13 +125,14 @@ module Sidekiq
       end
 
       # @api private
-      def perform(*arguments)
+      def perform(*args)
+        @_args = args.dup.freeze
         fetch_previous_iteration_state
 
         @_executions += 1
         @_start_time = ::Process.clock_gettime(::Process::CLOCK_MONOTONIC)
 
-        enumerator = build_enumerator(*arguments, cursor: @_cursor)
+        enumerator = build_enumerator(*args, cursor: @_cursor)
         unless enumerator
           logger.info("'#build_enumerator' returned nil, skipping the job.")
           return
@@ -112,7 +147,7 @@ module Sidekiq
         end
 
         completed = catch(:abort) do
-          iterate_with_enumerator(enumerator, arguments)
+          iterate_with_enumerator(enumerator, args)
         end
 
         on_stop
@@ -127,6 +162,10 @@ module Sidekiq
       end
 
       private
+
+      def is_cancelled?
+        @_cancelled = Sidekiq.redis { |c| c.hget("it-#{jid}", "cancelled") }
+      end
 
       def fetch_previous_iteration_state
         state = Sidekiq.redis { |conn| conn.hgetall(iteration_key) }
@@ -144,6 +183,12 @@ module Sidekiq
       STATE_TTL = 30 * 24 * 60 * 60 # one month
 
       def iterate_with_enumerator(enumerator, arguments)
+        if is_cancelled?
+          logger.info { "Job cancelled" }
+          return true
+        end
+
+        time_limit = Sidekiq.default_configuration[:timeout]
         found_record = false
         state_flushed_at = ::Process.clock_gettime(::Process::CLOCK_MONOTONIC)
 
@@ -153,14 +198,21 @@ module Sidekiq
 
           is_interrupted = interrupted?
           if ::Process.clock_gettime(::Process::CLOCK_MONOTONIC) - state_flushed_at >= STATE_FLUSH_INTERVAL || is_interrupted
-            flush_state
+            _, _, cancelled = flush_state
             state_flushed_at = ::Process.clock_gettime(::Process::CLOCK_MONOTONIC)
+            if cancelled == 1
+              @_cancelled = true
+              logger.info { "Job cancelled" }
+              return true
+            end
           end
 
           return false if is_interrupted
 
-          around_iteration do
-            each_iteration(object, *arguments)
+          verify_iteration_time(time_limit, object) do
+            around_iteration do
+              each_iteration(object, *arguments)
+            end
           end
         end
 
@@ -168,6 +220,16 @@ module Sidekiq
         true
       ensure
         @_runtime += (::Process.clock_gettime(::Process::CLOCK_MONOTONIC) - @_start_time)
+      end
+
+      def verify_iteration_time(time_limit, object)
+        start = ::Process.clock_gettime(::Process::CLOCK_MONOTONIC)
+        yield
+        finish = ::Process.clock_gettime(::Process::CLOCK_MONOTONIC)
+        total = finish - start
+        if total > time_limit
+          logger.warn { "Iteration took longer (%.2f) than Sidekiq's shutdown timeout (%d) when processing `%s`. This can lead to job processing problems during deploys" % [total, time_limit, object] }
+        end
       end
 
       def reenqueue_iteration_job
@@ -204,6 +266,7 @@ module Sidekiq
           conn.multi do |pipe|
             pipe.hset(key, state)
             pipe.expire(key, STATE_TTL)
+            pipe.hget(key, "cancelled")
           end
         end
       end